Add tooltips, notification, view detail

+++ src/views/hBpjsRates/List.vue
@@ -43,7 +43,9 @@
           <form>
             <div class="row mb-2">
               <div class="col-md-12">
-                <label>Description</label>
+                <label>Description 
+                  <font-awesome-icon :style="{ color: 'darkorange' }" v-c-tooltip.hover.click="'Required'" icon="info-circle"/>
+                </label>
                 <input :class="{ 'is-invalid' : errorMessages!=null && errorMessages.description!=undefined}" v-model="form.description" type="text" class="form-control">
                 <div class="invalid-feedback" v-if="errorMessages!=null && errorMessages.description!=undefined">
                     {{errorMessages.description[0]}}
@@ -52,14 +54,14 @@
             </div>
             <div class="row mb-2">
               <div class="col-md-6">
-                <label>JHTP</label>
+                <label>JHTP <font-awesome-icon :style="{ color: 'darkorange' }" v-c-tooltip.hover.click="'Required'" icon="info-circle"/></label>
                 <input :class="{ 'is-invalid' : errorMessages!=null && errorMessages.jhtp!=undefined}" v-model="form.jhtp" type="number" class="form-control">
                 <div class="invalid-feedback" v-if="errorMessages!=null && errorMessages.jhtp!=undefined">
                     {{errorMessages.jhtp[0]}}
                 </div>
               </div>
               <div class="col-md-6">
-                <label>JHTK</label>
+                <label>JHTK <font-awesome-icon :style="{ color: 'darkorange' }" v-c-tooltip.hover.click="'Required'" icon="info-circle"/></label>
                 <input :class="{ 'is-invalid' : errorMessages!=null && errorMessages.jhtk!=undefined}" v-model="form.jhtk" type="number" class="form-control">
                 <div class="invalid-feedback" v-if="errorMessages!=null && errorMessages.jhtk!=undefined">
                     {{errorMessages.jhtk[0]}}
@@ -68,14 +70,14 @@
             </div>
             <div class="row mb-2">
               <div class="col-md-6">
-                <label>JK</label>
+                <label>JK <font-awesome-icon :style="{ color: 'darkorange' }" v-c-tooltip.hover.click="'Required'" icon="info-circle"/></label>
                 <input :class="{ 'is-invalid' : errorMessages!=null && errorMessages.jk!=undefined}" v-model="form.jk" type="number" class="form-control">
                 <div class="invalid-feedback" v-if="errorMessages!=null && errorMessages.jk!=undefined">
                     {{errorMessages.jk[0]}}
                 </div>
               </div>
               <div class="col-md-6">
-                <label>JKK</label>
+                <label>JKK <font-awesome-icon :style="{ color: 'darkorange' }" v-c-tooltip.hover.click="'Required'" icon="info-circle"/></label>
                 <input :class="{ 'is-invalid' : errorMessages!=null && errorMessages.jkk!=undefined}" v-model="form.jkk" type="number" class="form-control">
                 <div class="invalid-feedback" v-if="errorMessages!=null && errorMessages.jkk!=undefined">
                     {{errorMessages.jkk[0]}}
@@ -84,14 +86,14 @@
             </div>
             <div class="row mb-2">
               <div class="col-md-6">
-                <label>JPK Lajang</label>
+                <label>JPK Lajang <font-awesome-icon :style="{ color: 'darkorange' }" v-c-tooltip.hover.click="'Required'" icon="info-circle"/></label>
                 <input :class="{ 'is-invalid' : errorMessages!=null && errorMessages.jpk_lajang!=undefined}" v-model="form.jpk_lajang" type="number" class="form-control">
                 <div class="invalid-feedback" v-if="errorMessages!=null && errorMessages.jpk_lajang!=undefined">
                     {{errorMessages.jpk_lajang[0]}}
                 </div>
               </div>
               <div class="col-md-6">
-                <label>JPK Nikah</label>
+                <label>JPK Nikah <font-awesome-icon :style="{ color: 'darkorange' }" v-c-tooltip.hover.click="'Required'" icon="info-circle"/></label>
                 <input :class="{ 'is-invalid' : errorMessages!=null && errorMessages.jpk_nikah!=undefined}" v-model="form.jpk_nikah" type="number" class="form-control">
                 <div class="invalid-feedback" v-if="errorMessages!=null && errorMessages.jpk_nikah!=undefined">
                     {{errorMessages.jpk_nikah[0]}}
@@ -100,14 +102,14 @@
             </div>
             <div class="row mb-2">
               <div class="col-md-6">
-                <label>BPJSP</label>
+                <label>BPJSP <font-awesome-icon :style="{ color: 'darkorange' }" v-c-tooltip.hover.click="'Required'" icon="info-circle"/></label>
                 <input :class="{ 'is-invalid' : errorMessages!=null && errorMessages.bpjsp!=undefined}" v-model="form.bpjsp" type="number" class="form-control">
                 <div class="invalid-feedback" v-if="errorMessages!=null && errorMessages.bpjsp!=undefined">
                     {{errorMessages.bpjsp[0]}}
                 </div>
               </div>
               <div class="col-md-6">
-                <label>BPJSK</label>
+                <label>BPJSK <font-awesome-icon :style="{ color: 'darkorange' }" v-c-tooltip.hover.click="'Required'" icon="info-circle"/></label>
                 <input :class="{ 'is-invalid' : errorMessages!=null && errorMessages.bpjsk!=undefined}" v-model="form.bpjsk" type="number" class="form-control">
                 <div class="invalid-feedback" v-if="errorMessages!=null && errorMessages.bpjsk!=undefined">
                     {{errorMessages.bpjsk[0]}}
@@ -116,14 +118,14 @@
             </div>
             <div class="row mb-2">
               <div class="col-md-6">
-                <label>Max Salary Pension</label>
+                <label>Max Salary Pension <font-awesome-icon :style="{ color: 'darkorange' }" v-c-tooltip.hover.click="'Required'" icon="info-circle"/></label>
                 <input :class="{ 'is-invalid' : errorMessages!=null && errorMessages.max_salary_pension!=undefined}" v-model="form.max_salary_pension" type="text" class="form-control">
                 <div class="invalid-feedback" v-if="errorMessages!=null && errorMessages.max_salary_pension!=undefined">
                     {{errorMessages.max_salary_pension[0]}}
                 </div>
               </div>
               <div class="col-md-6">
-                <label>Max Salary Medical</label>
+                <label>Max Salary Medical <font-awesome-icon :style="{ color: 'darkorange' }" v-c-tooltip.hover.click="'Required'" icon="info-circle"/></label>
                 <input :class="{ 'is-invalid' : errorMessages!=null && errorMessages.max_salary_medical!=undefined}" v-model="form.max_salary_medical" type="text" class="form-control">
                 <div class="invalid-feedback" v-if="errorMessages!=null && errorMessages.max_salary_medical!=undefined">
                     {{errorMessages.max_salary_medical[0]}}
@@ -133,14 +135,14 @@
             
             <div class="row mb-2">
               <div class="col-md-6">
-                <label>Pension Company</label>
+                <label>Pension Company <font-awesome-icon :style="{ color: 'darkorange' }" v-c-tooltip.hover.click="'Required'" icon="info-circle"/></label>
                 <input :class="{ 'is-invalid' : errorMessages!=null && errorMessages.pension_company!=undefined}" v-model="form.pension_company" type="number" class="form-control">
                 <div class="invalid-feedback" v-if="errorMessages!=null && errorMessages.pension_company!=undefined">
                     {{errorMessages.pension_company[0]}}
                 </div>
               </div>
               <div class="col-md-6">
-                <label>Pension Employees</label>
+                <label>Pension Employees <font-awesome-icon :style="{ color: 'darkorange' }" v-c-tooltip.hover.click="'Required'" icon="info-circle"/></label>
                 <input :class="{ 'is-invalid' : errorMessages!=null && errorMessages.pension_employees!=undefined}" v-model="form.pension_employees" type="number" class="form-control">
                 <div class="invalid-feedback" v-if="errorMessages!=null && errorMessages.pension_employees!=undefined">
                     {{errorMessages.pension_employees[0]}}
@@ -155,6 +157,85 @@
         </template>
       </b-modal>
     </div>
+    <!--Modal for CRUD end here --> 
+    <!--Modal for View -->
+    <b-modal 
+        id="view"      
+        no-close-on-backdrop
+        size = "md"  
+        title="View Detail" 
+    >
+      <div class="row mb-2">
+        <div class="col-md-12">
+          <label>Description</label>
+          <div class="data-view">{{form.description}}</div>
+        </div>
+      </div>
+      <div class="row mb-2">
+        <div class="col-md-6">
+          <label>JHTP</label>
+          <div class="data-view">{{form.jhtp}}</div>
+        </div>
+        <div class="col-md-6">
+          <label>JHTK</label>
+          <div class="data-view">{{form.jhtk}}</div>
+        </div>
+      </div>
+      <div class="row mb-2">
+        <div class="col-md-6">
+          <label>JK</label>
+          <div class="data-view">{{form.jk}}</div>
+        </div>
+        <div class="col-md-6">
+          <label>JKK</label>
+          <div class="data-view">{{form.jkk}}</div>
+        </div>
+      </div>
+      <div class="row mb-2">
+        <div class="col-md-6">
+          <label>JPK Lajang</label>
+          <div class="data-view">{{form.jpk_lajang}}</div>
+        </div>
+        <div class="col-md-6">
+          <label>JPK Nikah</label>
+          <div class="data-view">{{form.jpk_nikah}}</div>
+        </div>
+      </div>
+      <div class="row mb-2">
+        <div class="col-md-6">
+          <label>BPJSP</label>
+          <div class="data-view">{{form.bpjsp}}</div>
+        </div>
+        <div class="col-md-6">
+          <label>BPJSK</label>
+          <div class="data-view">{{form.bpjsk}}</div>
+        </div>
+      </div>
+      <div class="row mb-2">
+        <div class="col-md-6">
+          <label>Max Salary Pension</label>
+          <div class="data-view">{{form.max_salary_pension}}</div>
+        </div>
+        <div class="col-md-6">
+          <label>Max Salary Medical</label>
+          <div class="data-view">{{form.max_salary_medical}}</div>
+        </div>
+      </div>          
+      <div class="row mb-2">
+        <div class="col-md-6">
+          <label>Pension Company</label>
+          <div class="data-view">{{form.pension_company}}</div>
+        </div>
+        <div class="col-md-6">
+          <label>Pension Employees</label>
+          <div class="data-view">{{form.pension_employees}}</div>
+        </div>
+      </div>
+      <template v-slot:modal-footer="{close}">
+        <ButtonCloseFull ref="btnClose" :actions="closeActView" />
+      </template>
+    </b-modal>
+    <!--Modal View end here -->
     
   </CRow>
 </template>
@@ -168,7 +249,11 @@ import ButtonAddFull from '../../components/button/ButtonAddFull.vue'
 import ButtonSaveFull from '../../components/button/ButtonSaveFull.vue'
 import ButtonCloseFull from '../../components/button/ButtonCloseFull.vue'
 
+//axios
 import axios from 'axios'
+
+//convert helper
+import {convertToMoney} from '../../helpers/convertion'
 
 const url = process.env.VUE_APP_API_SOURCE;
 export default {
@@ -280,22 +365,27 @@ export default {
       this.$bvModal.show('add-edit');    
     },
     deleteAct(data){
-
-      axios.delete(''+url+'bpjsrate/'+data.code+'',{
-          headers : {
-            'Authorization' : ''+this.$store.getters.curentCompany.token_type+' '+this.$store.getters.curentCompany.access_token+'',
-            'tenant-token'  : ''+this.$store.getters.curentUser.api_token+''
-          }
+      if (confirm("Are you sure delete selected data?")==true){
+        this.status = "delete";
+        axios.delete(''+url+'bpjsrate/'+data.code+'',{
+            headers : {
+              'Authorization' : ''+this.$store.getters.curentCompany.token_type+' '+this.$store.getters.curentCompany.access_token+'',
+              'tenant-token'  : ''+this.$store.getters.curentUser.api_token+''
+            }
         }).then((res)=>{
             this.isLoading = false;
-            this.processResponse(res);
+            this.processResponse(res,this.status);
         }).catch((err)=>{
             this.isLoading = false;
             console.log(err);
         })
+      }
     },
     viewAct(data){
-      alert(JSON.stringify(data)); 
+      this.form = Object.assign({},data);
+      this.form.max_salary_pension =convertToMoney(this.form.max_salary_pension);
+      this.form.max_salary_medical =convertToMoney(this.form.max_salary_medical);
+      this.$bvModal.show('view');
     },
     newAct(){
       this.titleModal = "Add New Data";
@@ -305,7 +395,7 @@ export default {
       this.errorMessages = null
       this.$bvModal.show('add-edit');
     },
-    processResponse(res){
+    processResponse(res,status){
       this.errorMessage = '';
       this.errorMessages = null
       switch(res.data.message){
@@ -320,6 +410,30 @@ export default {
           this.isError = false;
           this.$refs.gridajax.refreshTable();
           this.$bvModal.hide('add-edit');
+
+          //toasted
+          let messageT = "";
+         
+          switch (status){
+            case "save" :
+              messageT = "Data saved successfully";
+            break;
+            case "update" :
+              messageT = "Data updated successfully";
+            break;
+            case "delete" :
+              messageT = "Data deleted successfully";
+            break;
+          }
+
+          this.$toasted.show(messageT,{
+            'position' : 'bottom-right',           
+            'type' : 'success',
+            'theme' : 'bubble',   
+            'duration' : 5000,        
+            'closeOnSwipe' : true
+          });
+
         break;
       }
     },
@@ -335,7 +449,7 @@ export default {
           }
         }).then((res)=>{
             this.isLoading = false;
-            this.processResponse(res);
+            this.processResponse(res,this.status);
         }).catch((err)=>{
             this.isLoading = false;
             console.log(err);
@@ -348,7 +462,7 @@ export default {
           }
         }).then((res)=>{
             this.isLoading = false;
-            this.processResponse(res);
+            this.processResponse(res,this.status);
         }).catch((err)=>{
             this.isLoading = false;
             console.log(err);
@@ -358,6 +472,9 @@ export default {
     closeAct(){
       this.$bvModal.hide('add-edit');
     },
+    closeActView(){
+      this.$bvModal.hide('view');
+    }
     
   }
   
@@ -369,5 +486,10 @@ export default {
   }
   .dis{
     opacity: 0.8;
+  }
+
+  .data-view{
+    border-bottom: solid thin #e5e5e5;
+    padding: 5px;
   }
 </style>
